repack: avoid loosening promisor pack objects in partial clones
When `git repack -A -d` is run in a partial clone, `pack-objects` is
invoked twice: once to repack all promisor objects, and once to repack
all non-promisor objects. The latter `pack-objects` invocation is with
--exclude-promisor-objects and --unpack-unreachable, which loosens all
unused objects. Unfortunately, this includes promisor objects.

Because the -d argument to `git repack` subsequently deletes all loose
objects also in packs, these just-loosened promisor objects will be
immediately deleted. However, this extra disk churn is unnecessary in
the first place. For example, a freshly-cloned partial repo that
filters all blob objects (e.g. `--filter=blob:none`), `repack` ends up
unpacking all promisor objects (trees, commits) into the filesystem
that, depending o the repo size, increases the disk usage considerably.
In my copy of the linux.git, the object directory peaked 26GB of
increased disk usage.

In order to avoid this extra disk churn, pass the names of the
promisor packfiles as --keep-pack arguments to the second invocation of
`pack-objects`. This informs `pack-objects` that the promisor objects
are already in a safe packfile and, therefore, do not need to be
loosened. The --keep-pack option takes only a packfile name, but we
concatenate both the path and the name in a single string. Instead,
let's split them into separate string in order to easily pass the
packfile name later.

As the "evidence", loosened objects, are deleted by the process, this
prevents inspecting the object directory in order to test the behavior.
So, teach `git pack-objects` to count all the loosened objects and emit
to trace2 that is used on the added regression test.

Finally, add a new perf test to evaluate the performance impact
made by this changes (tested on git.git):

     Test          HEAD^                 HEAD
     ----------------------------------------------------------
     5600.3: gc    134.38(41.93+90.95)   7.80(6.72+1.35) -94.2%

For a bigger repository, such as linux.git, the improvement is even
more considerably:

     Test          HEAD^                     HEAD
     -------------------------------------------------------------------
     5600.3: gc    6833.00(918.07+3162.74)   268.79(227.02+39.18) -96.1%

These improvements are particular big because every object in the
newly-cloned partial repository is a promisor object.

Reported-by: SZEDER Gábor <szeder.dev@gmail.com>
Helped-by: Jeff King <peff@peff.net>
Helped-by: Jonathan Tan <jonathantanmy@google.com>
Signed-off-by: Rafael Silva <rafaeloliveira.cs@gmail.com>